remaining app services converted to machete-client; consolidated reports
