+++ src/app/reports/models/report.ts
@@ -1,9 +1,10 @@
+import { ReportDefinitionVM } from "machete-client";
 import { Column } from "./column";
 import { SearchInputs } from "./search-inputs";
 /**
  * Created by jcarter on 3/9/17.
  */
-export class Report {
+export class Report implements ReportDefinitionVM {
   constructor(init?: Partial<Report>) {
     Object.assign(this, init);
   }
@@ -16,10 +17,8 @@ export class Report {
   subcategory: string;
   idString: string;
   id: number;
-  // eslint-disable-next-line @typescript-eslint/ban-types
-  data: Object;
-  datecreated: Date;
-  dateupdated: Date;
+  datecreated: string;
+  dateupdated: string;
   createdby: string;
   updatedby: string;
   idPrefix: string;
